requirements.txt
fastapi~=0.120.2
uvicorn
pyyaml~=6.0.3
python-multipart
sentence-transformers~=5.1.2
scikit-learn~=1.7.2

numpy~=2.3.4
faker==24.0.0
locust~=2.42.1
aiohttp
pandas
langgraph
langchain_core
langchain
langchain_google_genai
dotenv
streamlit